// src/app/model/enum/solution-status.enum.ts
export enum SolutionStatus {
    MEMORIZE_PERIOD = 'Memorize the potential answers before they flip.',
    SOLVE_PERIOD = 'Solve and select the correct answer.',
    SOLUTION_FOUND = 'You got the correct answer! Next round starting soon.',
    SOLUTION_NOT_FOUND = 'You got the wrong answer. Next round starting soon.',
    SOLUTION_NOT_SELECTED = 'Time ran out before an answer was selected. Next round starting soon.'
}